Add files via upload
This script generates rebalancing signals for a portfolio of global assets that I call a 'weatherproof' portfolio. The rebalancing rule: On a monthly basis, invest in the top 5 out of 10 assets, ranked based on the average of their 12, 6, 3, and 1 month total return, if the price of the asset is above its 9 month moving average. The capital that would be allocated to the bottom 5 assets is instead put in cash/T-bills (BIL). The signals from this script determine which assets are allocated to each month.